+++ src/core/decrypt/utils/persistent_sessions_store.ts
@@ -16,6 +16,7 @@
 
 import { ICustomMediaKeySession } from "../../../compat";
 import log from "../../../log";
+import { IPersistentSessionStorage } from "../../../public_types";
 import areArraysOfNumbersEqual from "../../../utils/are_arrays_of_numbers_equal";
 import { assertInterface } from "../../../utils/assert";
 import { bytesToBase64 } from "../../../utils/base64";
@@ -23,9 +24,8 @@ import hashBuffer from "../../../utils/hash_buffer";
 import isNonEmptyString from "../../../utils/is_non_empty_string";
 import isNullOrUndefined from "../../../utils/is_null_or_undefined";
 import {
-  IProcessedProtectionData,
   IPersistentSessionInfo,
-  IPersistentSessionStorage,
+  IProcessedProtectionData,
 } from "../types";
 import areInitializationValuesCompatible from "./are_init_values_compatible";
 import { IFormattedInitDataValue } from "./init_data_values_container";
