+++ src/core/init/stream_events_emitter/index.ts
@@ -19,14 +19,12 @@ import {
   IPublicNonFiniteStreamEvent,
   IPublicStreamEvent,
   IStreamEvent,
-  IStreamEventData,
   IStreamEventEvent,
   IStreamEventSkipEvent,
 } from "./types";
 
 export {
   IStreamEvent,
-  IStreamEventData,
   IPublicNonFiniteStreamEvent,
   IPublicStreamEvent,
