+++ src/manifest/representation.ts
@@ -20,11 +20,11 @@ import {
   IContentProtections,
   IParsedRepresentation,
 } from "../parsers/manifest";
+import { IHDRInformation } from "../public_types";
 import areArraysOfNumbersEqual from "../utils/are_arrays_of_numbers_equal";
 import { IRepresentationIndex } from "./representation_index";
 import {
   IAdaptationType,
-  IHDRInformation,
 } from "./types";
 
 /**
